python/fastapi-do-zero/fast_zero/pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["humbertoluis22 <humbertoluis2013@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pytest = "^8.3.4"
funcsigs = "^1.0.2"
scikit-learn = "^1.6.1"
sqlalchemy = "^2.0.37"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
tzdata = "^2025.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style= 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
run = 'fastapi dev fast_zero/app.py'

test =  'pytest --cov=fast_zero -vv'
post_test = 'coverage html'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . -- fiz && ruff format'
